+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to the z/OS FTP Plug-in for Zowe CLI will be documented in this file.
 
 ## Recent Changes
+
+- Support listing loadlib members
+- Clear password error message
 - Support listing jobs without default prefix *.
 
 ## `1.6.0`